pyproject.toml
[project]
name = "biomol_surface_unsup"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
  "torch",
  "numpy",
  "pyyaml",
  "scikit-image",
  "tqdm",
]

[tool.pytest.ini_options]
pythonpath = ["src"]